+++ src/layouts/BaseLayout.tsx
@@ -1,0 +1,5 @@
+const BasicLayout = () => {
+  return <div>BasicLayout</div>;
+};
+
+export default BasicLayout;