+++ src/pages/about/index.tsx
@@ -1,0 +1,5 @@
+const About = () => {
+  return <div>我是关于我们</div>;
+};
+
+export default About;