+++ src/pages/error/NoFund404.tsx
@@ -1,0 +1,4 @@
+const NoFund404 = () => {
+  return <div>404</div>;
+};
+export default NoFund404;