+++ src/pages/home/index.tsx
@@ -1,0 +1,5 @@
+const Home = () => {
+  return <div>我是主页</div>;
+};
+
+export default Home;